Clean up a lot of things from helper-zcl that don't belong there.
